memindahkan halaman error 404 ke not found, dan menambahkan logika di tombol agar berpindah halaman.
