--- app/(public)/customer/error-404/page.tsx
@@ -1,26 +1,0 @@
-export default function NotFoundSection() {
-  return (
-    <div className="flex flex-col items-center justify-center text-center min-h-screen bg-white px-4 font-sans">
-      <img
-        src="https://cdn-icons-png.flaticon.com/512/7486/7486758.png"
-        alt="Error"
-        className="w-72 mb-6"
-      />
-      <h2 className="text-2xl md:text-3xl font-semibold text-gray-500 mb-2">
-        404, Halaman Tidak Ditemukan
-      </h2>
-      <p className="text-gray-500 max-w-lg mb-8 leading-relaxed">
-        Terjadi kesalahan. Tampaknya permintaan Anda tidak dapat ditemukan.
-        Kemungkinan tautannya rusak atau halamannya telah dihapus.
-      </p>
-      <div className="flex space-x-4">
-        <button className="bg-orange-500 hover:bg-orange-600 text-white font-medium px-6 py-2 rounded-md transition-colors">
-          ← KEMBALI
-        </button>
-        <button className="bg-orange-100 hover:bg-orange-200 text-orange-600 border border-orange-400 font-medium px-6 py-2 rounded-md transition-colors">
-          ⟳ KEMBALI KE BERANDA
-        </button>
-      </div>
-    </div>
-  );
-}
