+++ app/(public)/not-found.tsx
@@ -1,211 +1,47 @@
-import { EventButton } from "@/components/checkout/cart/event-button";
+'use client';
 
-export default function NotFound() {
+import { useRouter } from 'next/navigation';
+
+export default function NotFoundSection() {
+  const router = useRouter();
+
+  const handleGoBack = () => {
+    router.back();
+  };
+
+  const handleGoHome = () => {
+    router.push('/');
+  };
+
   return (
-    <div className="justify-cente my-0 flex flex-col flex-wrap items-center md:my-12">
-      <div className="h-fit">
-        <svg
-          className="size-64 md:size-96 fill-white stroke-black dark:fill-neutral-900 dark:stroke-white"
-          fill="none"
-          viewBox="0 0 850 500"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M469.391 25.5H614.391C696.405 25.5 762.891 91.9857 762.891 174V305.5H426.891V68C426.891 44.5279 445.919 25.5 469.391 25.5Z"
-            strokeWidth={3}
-          />
-          <path
-            d="M730.657 235.073C736.465 238.426 740.524 242.271 739.722 243.66C738.92 245.049 733.561 243.457 727.752 240.103C721.943 236.749 716.584 235.156 715.782 236.545C714.98 237.935 719.039 241.779 724.848 245.133"
-            strokeWidth={2}
-          />
-          <path
-            d="M727.752 240.103C733.561 243.457 737.62 247.301 736.818 248.69C736.016 250.079 730.657 248.487 724.848 245.133C719.039 241.779 713.68 240.187 712.878 241.576C712.076 242.965 716.135 246.81 721.944 250.163"
-            strokeWidth={2}
-          />
-          <path
-            d="M724.848 245.133C730.657 248.487 734.716 252.332 733.914 253.721C733.112 255.11 727.753 253.517 721.944 250.163C716.135 246.81 710.776 245.217 709.974 246.606C709.172 247.995 713.231 251.84 719.039 255.194"
-            strokeWidth={2}
-          />
-          <path
-            d="M721.944 250.163C727.753 253.517 731.811 257.362 731.009 258.751C730.207 260.14 724.848 258.547 719.039 255.194C713.231 251.84 707.871 250.247 707.069 251.636C706.267 253.025 710.326 256.87 716.135 260.224"
-            strokeWidth={2}
-          />
-          <path
-            d="M719.039 255.194C724.848 258.547 728.907 262.392 728.105 263.781C727.303 265.17 721.944 263.578 716.135 260.224C710.326 256.87 704.967 255.278 704.165 256.667C703.363 258.056 707.422 261.9 713.231 265.254"
-            strokeWidth={2}
-          />
-          <path
-            d="M716.135 260.224C721.944 263.578 726.003 267.422 725.201 268.812C724.399 270.201 719.04 268.608 713.231 265.254C707.422 261.9 702.063 260.308 701.261 261.697C700.459 263.086 704.518 266.931 710.327 270.285"
-            strokeWidth={2}
-          />
-          <path
-            d="M713.231 265.254C719.04 268.608 723.099 272.453 722.297 273.842C721.495 275.231 716.135 273.638 710.327 270.284C704.518 266.931 699.159 265.338 698.357 266.727C697.555 268.116 701.613 271.961 707.422 275.315"
-            strokeWidth={2}
-          />
-          <path
-            d="M710.326 270.285C716.135 273.638 720.194 277.483 719.392 278.872C718.59 280.261 713.231 278.669 707.422 275.315C701.613 271.961 696.254 270.368 695.452 271.757C694.65 273.147 698.709 276.991 704.518 280.345"
-            strokeWidth={2}
-          />
-          <path
-            d="M389.986 25.5C410.688 25.5 426.891 38.9579 426.891 54.917V272.5H79.8906V129.398C79.8906 113.492 84.3455 98.4124 92.3252 84.917L107.57 97.1904L107.982 97.5225H137.515L137.856 96.4961L144.836 75.5254L165.173 71.1387L167.408 70.6562L166.075 68.7979L152.672 50.1094L176.441 39.0615L176.607 38.9854L176.75 38.8701L193.357 25.5H389.986Z"
-            strokeWidth={3}
-          />
-          <path
-            d="M105.732 227.042C110.555 224.257 115.005 222.935 115.67 224.088C116.336 225.241 112.966 228.434 108.144 231.218C103.321 234.002 99.951 237.195 100.617 238.348C101.283 239.501 105.732 238.179 110.555 235.394"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M108.144 231.218C112.967 228.434 117.417 227.111 118.082 228.264C118.748 229.418 115.378 232.61 110.556 235.394C105.733 238.179 102.363 241.371 103.029 242.524C103.695 243.677 108.144 242.355 112.967 239.571"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M110.555 235.394C115.377 232.61 119.827 231.287 120.493 232.441C121.158 233.594 117.789 236.786 112.966 239.571C108.143 242.355 104.773 245.547 105.439 246.7C106.105 247.854 110.554 246.531 115.377 243.747"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M112.967 239.571C117.789 236.786 122.239 235.464 122.905 236.617C123.571 237.77 120.201 240.963 115.378 243.747C110.555 246.531 107.185 249.723 107.851 250.877C108.517 252.03 112.966 250.708 117.789 247.923"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M115.379 243.747C120.202 240.963 124.651 239.64 125.317 240.793C125.983 241.947 122.613 245.139 117.79 247.923C112.967 250.708 109.597 253.9 110.263 255.053C110.929 256.206 115.379 254.884 120.201 252.1"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M117.789 247.923C122.612 245.139 127.061 243.817 127.727 244.97C128.393 246.123 125.023 249.315 120.2 252.1C115.377 254.884 112.008 258.076 112.673 259.229C113.339 260.383 117.789 259.06 122.611 256.276"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M120.201 252.1C125.024 249.315 129.473 247.993 130.139 249.146C130.805 250.299 127.435 253.492 122.612 256.276C117.79 259.06 114.42 262.252 115.086 263.406C115.751 264.559 120.201 263.237 125.024 260.452"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M122.611 256.276C127.434 253.492 131.883 252.169 132.549 253.323C133.215 254.476 129.845 257.668 125.022 260.452C120.2 263.237 116.83 266.429 117.496 267.582C118.162 268.735 122.611 267.413 127.434 264.629"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M395.732 148.042C400.555 145.257 405.005 143.935 405.67 145.088C406.336 146.241 402.966 149.434 398.144 152.218C393.321 155.002 389.951 158.195 390.617 159.348C391.283 160.501 395.732 159.179 400.555 156.394"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M398.144 152.218C402.967 149.434 407.417 148.111 408.082 149.264C408.748 150.418 405.378 153.61 400.556 156.394C395.733 159.179 392.363 162.371 393.029 163.524C393.695 164.677 398.144 163.355 402.967 160.571"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M400.555 156.394C405.377 153.61 409.827 152.288 410.493 153.441C411.158 154.594 407.789 157.786 402.966 160.571C398.143 163.355 394.773 166.547 395.439 167.7C396.105 168.854 400.554 167.531 405.377 164.747"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M402.967 160.571C407.789 157.786 412.239 156.464 412.905 157.617C413.571 158.77 410.201 161.963 405.378 164.747C400.555 167.531 397.185 170.724 397.851 171.877C398.517 173.03 402.966 171.708 407.789 168.923"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M405.377 164.747C410.2 161.963 414.649 160.64 415.315 161.794C415.981 162.947 412.611 166.139 407.788 168.923C402.965 171.708 399.596 174.9 400.261 176.053C400.927 177.206 405.377 175.884 410.199 173.1"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M407.789 168.923C412.612 166.139 417.061 164.817 417.727 165.97C418.393 167.123 415.023 170.315 410.2 173.1C405.377 175.884 402.008 179.076 402.673 180.229C403.339 181.383 407.789 180.06 412.611 177.276"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M410.199 173.1C415.022 170.315 419.471 168.993 420.137 170.146C420.803 171.299 417.433 174.492 412.61 177.276C407.788 180.06 404.418 183.253 405.084 184.406C405.749 185.559 410.199 184.237 415.022 181.452"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M412.611 177.276C417.434 174.492 421.883 173.169 422.549 174.323C423.215 175.476 419.845 178.668 415.022 181.452C410.2 184.237 406.83 187.429 407.496 188.582C408.162 189.735 412.611 188.413 417.434 185.629"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M55.8906 460.5V466C55.8906 467.381 54.7713 468.5 53.3906 468.5H50.3906C49.0099 468.5 47.8906 467.381 47.8906 466V460.5H55.8906Z"
-            fill="#FCFCFC"
-            strokeWidth={3}
-          />
-          <path
-            d="M582.939 146.493C582.939 146.493 536.267 128.734 517.204 123.996C521.493 135.367 517.204 202.17 508.626 212.119C500.047 222.069 490.039 271.815 495.758 281.765C501.477 291.714 557.236 274.658 582.971 271.815C603.558 269.541 649.69 277.501 670.183 281.765C659.889 268.12 674.472 175.638 683.05 131.103C676.188 125.418 582.939 146.493 582.939 146.493Z"
-            strokeWidth={3}
-          />
-          <path d="M670.877 281.237C671.691 280.656 672.505 280.075 673.319 279.494C672.661 278.616 671.992 277.689 671.34 276.76C660.21 260.011 649.291 241.225 649.634 220.479C649.713 219.343 649.852 218.203 650.058 217.09C649.803 218.193 649.614 219.329 649.485 220.466C648.227 241.348 658.395 260.82 668.972 278.346C669.596 279.323 670.24 280.302 670.877 281.237Z" />
-          <path d="M493.336 270.788C493.917 271.601 494.498 272.415 495.079 273.229C495.799 272.685 496.526 272.132 497.248 271.582C509.904 261.809 524.376 251.848 541.059 253.714C541.968 253.812 542.849 253.938 543.754 254.1C542.86 253.889 541.985 253.713 541.079 253.565C524.448 250.707 508.775 260.159 495.586 269.267C494.834 269.776 494.08 270.288 493.336 270.788Z" />
-          <path d="M495.175 273.536C494.531 272.771 493.886 272.006 493.242 271.242C493.565 271.002 493.899 270.744 494.228 270.483C499.959 265.696 505.846 260.142 508.005 252.69C508.103 252.274 508.181 251.85 508.232 251.426C508.231 251.853 508.203 252.286 508.153 252.716C506.842 260.532 501.516 266.931 496.129 272.607C495.815 272.921 495.492 273.236 495.175 273.536Z" />
-          <path d="M618.585 154.387L602.231 215.419L614.279 218.647L611.225 230.044L599.177 226.816L593.172 249.228L578.458 245.285L584.463 222.873L541.233 211.29L541.446 210.498L541.587 210.394L617.533 154.3L617.727 154.157L618.585 154.387ZM566.43 205.826L587.517 211.476L594.798 184.301L566.43 205.826Z" />
-          <path
-            d="M182.42 229.381C167.566 209.856 157.597 184.793 154.469 174.703C156.124 167.193 238.277 146.308 281.347 143.492C315.802 141.239 333.213 129.569 337.611 124.015C334.971 129.084 339.163 142.084 341.589 147.951L326.16 282.182L203.043 318.555L182.42 229.381Z"
-            strokeWidth={3}
-          />
-          <path
-            d="M396.733 171.243C382.046 157.128 338.78 147.819 318.984 144.929C314.304 148.797 319.658 196.882 325.538 224.84C331.419 252.799 322.675 283.656 324.821 291.071C326.966 298.487 385.464 286.892 404.211 285.836C422.958 284.78 480.957 307.728 486.779 306.854C492.601 305.98 489.225 252.322 483.791 234.497C479.444 220.236 481.584 173.746 483.198 152.283C466.149 150.622 418.451 164.231 396.733 171.243Z"
-            strokeWidth={3}
-          />
-          <path d="M252.42 170.288L278.557 227.814L289.912 222.655L294.793 233.397L283.437 238.556L293.035 259.681L279.166 265.982L269.568 244.858L228.822 263.37L228.483 262.624L228.526 262.454L251.553 170.889L251.612 170.655L252.42 170.288ZM244.812 243.146L264.688 234.116L253.05 208.502L244.812 243.146Z" />
-          <path d="M396.332 184.169C400.609 183.899 404.573 184.586 408.213 186.237L408.903 186.557C412.323 188.204 415.322 190.629 417.898 193.823L417.897 193.824C420.706 197.144 422.96 201.259 424.665 206.157L424.664 206.157C426.436 210.981 427.517 216.521 427.912 222.773C428.312 229.105 427.941 234.818 426.795 239.905L426.796 239.906C425.72 244.984 423.999 249.353 421.627 253.003C419.325 256.646 416.44 259.521 412.973 261.618L412.968 261.62C409.56 263.635 405.713 264.774 401.438 265.044C397.16 265.314 393.162 264.629 389.455 262.983C385.75 261.338 382.492 258.851 379.681 255.531L379.674 255.522C376.93 252.119 374.708 247.962 373.002 243.063C371.295 238.157 370.249 232.614 369.859 226.439C369.461 220.133 369.689 214.548 370.549 209.69L370.729 208.729C371.73 203.577 373.38 199.209 375.691 195.639L375.695 195.634L375.696 195.634C378.074 192.065 381.006 189.345 384.496 187.486C387.976 185.55 391.925 184.448 396.332 184.169ZM396.993 196.294C393.109 196.54 389.967 197.94 387.531 200.481L387.526 200.487C385.067 202.97 383.281 206.36 382.191 210.687C381.169 215.014 380.833 220.024 381.193 225.723C381.547 231.342 382.511 236.271 384.075 240.516L384.387 241.284C385.978 245.066 388.075 247.992 390.664 250.088C393.393 252.222 396.684 253.178 400.57 252.932C404.459 252.687 407.606 251.323 410.046 248.859C412.569 246.292 414.383 242.82 415.468 238.416C416.554 234.004 416.922 228.951 416.562 223.25C416.202 217.549 415.207 212.665 413.586 208.587L413.585 208.585C411.961 204.432 409.73 201.298 406.912 199.147L406.907 199.143C404.176 197.005 400.882 196.049 396.993 196.294Z" />
-          <path d="M484.247 228.109C484.151 227.113 484.054 226.118 483.958 225.123C483.984 225.455 483.993 225.621 483.986 225.622C483.381 225.681 482.769 225.742 482.16 225.802C477.39 226.281 472.645 226.703 467.941 226.442C461.772 226.119 455.611 224.799 450.29 221.354C449.766 221.017 449.265 220.675 448.758 220.302C449.186 220.763 449.618 221.196 450.082 221.623C454.78 226.031 461.426 228.135 467.831 228.439C472.725 228.709 477.58 228.271 482.358 227.792C482.967 227.732 483.578 227.671 484.179 227.613C484.186 227.612 484.208 227.777 484.247 228.109Z" />
-          <path d="M154.816 172.874C154.402 173.784 153.987 174.694 153.573 175.604C153.717 175.303 153.792 175.154 153.798 175.157C154.352 175.409 154.913 175.662 155.47 175.914C159.84 177.884 164.161 179.892 168.104 182.47C173.285 185.834 177.96 190.057 180.846 195.702C181.131 196.255 181.394 196.803 181.647 197.379C181.507 196.765 181.349 196.174 181.161 195.572C179.296 189.406 174.593 184.261 169.198 180.795C165.094 178.114 160.671 176.066 156.294 174.092C155.736 173.84 155.177 173.587 154.627 173.337C154.621 173.334 154.684 173.18 154.816 172.874Z" />
-          <path d="M482.854 229.286C482.273 228.472 481.692 227.658 481.111 226.844C481.299 227.119 481.391 227.259 481.385 227.263C480.89 227.616 480.391 227.975 479.894 228.332C476.003 231.131 472.104 233.869 467.9 235.996C462.396 238.8 456.4 240.737 450.069 240.414C449.448 240.385 448.842 240.338 448.217 240.269C448.818 240.455 449.408 240.614 450.024 240.752C456.296 242.219 463.104 240.719 468.803 237.78C473.177 235.567 477.162 232.76 481.06 229.956C481.557 229.599 482.056 229.241 482.547 228.89C482.553 228.886 482.655 229.018 482.854 229.286Z" />
-          <path d="M156.611 172.551C156.707 173.546 156.803 174.541 156.899 175.537C156.874 175.204 156.864 175.038 156.871 175.037C157.477 174.979 158.089 174.918 158.697 174.857C163.467 174.378 168.212 173.957 172.916 174.217C179.085 174.54 185.246 175.86 190.567 179.305C191.091 179.642 191.592 179.985 192.099 180.357C191.671 179.896 191.239 179.463 190.775 179.036C186.077 174.629 179.432 172.524 173.027 172.22C168.132 171.95 163.277 172.388 158.5 172.867C157.89 172.927 157.28 172.988 156.678 173.047C156.672 173.047 156.649 172.882 156.611 172.551Z" />
-          <path
-            d="M802.67 465.307V472C802.67 473.38 801.55 474.499 800.17 474.5H797.719C796.338 474.5 795.219 473.38 795.219 472V465.307H802.67Z"
-            fill="#FCFCFC"
-            strokeWidth={3}
-          />
-          <path
-            d="M674.476 259.033C678.066 257.051 683.625 255.09 688.676 255.465C691.17 255.65 693.483 256.4 695.399 257.939C697.31 259.472 698.944 261.884 699.928 265.604L700.104 266.268L700.721 266.568C703.604 267.974 707.579 270.619 710.185 273.93C712.76 277.203 713.836 280.861 711.674 284.712L714.287 286.186C716.369 282.51 719.832 277.14 723.477 272.913C725.304 270.794 727.122 269.026 728.788 267.894C730.508 266.725 731.753 266.443 732.581 266.668C734.471 267.181 736.512 266.877 738.454 266.207C740.411 265.532 742.406 264.439 744.285 263.191C747.782 260.867 751.044 257.895 753.153 255.691H754.035C776.452 255.691 794.129 274.766 792.424 297.118L787.196 365.637H47.7793L39.6406 315.431C34.5609 284.094 58.7389 255.805 90.3828 257.061C142.397 259.124 198.116 263.163 250.171 266.688C302.203 270.211 350.593 273.223 387.877 273.223C388.95 273.223 390.472 273.683 392.282 274.529C394.05 275.356 395.93 276.473 397.668 277.61C399.401 278.745 400.965 279.882 402.098 280.737C402.663 281.164 403.119 281.52 403.433 281.768C403.589 281.891 403.71 281.988 403.791 282.053C403.831 282.085 403.862 282.11 403.882 282.126C403.892 282.134 403.9 282.141 403.904 282.145C403.907 282.146 403.908 282.148 403.909 282.149L403.91 282.15L404.96 283.008L405.92 282.05V282.049C405.921 282.048 405.923 282.047 405.925 282.045C405.929 282.041 405.935 282.035 405.942 282.027C405.959 282.011 405.984 281.987 406.017 281.954C406.083 281.889 406.181 281.793 406.309 281.67C406.564 281.423 406.937 281.069 407.399 280.644C408.325 279.791 409.603 278.658 411.019 277.528C412.44 276.394 413.967 275.288 415.396 274.474C416.881 273.628 418.045 273.223 418.795 273.223C457.11 273.223 479.701 268.819 504.804 264.435C529.662 260.094 557.066 255.758 604.938 255.693L609.761 270.536L610.374 272.424L612.022 271.319C616.909 268.046 624.769 264.145 632.442 263.016C636.267 262.453 639.957 262.592 643.188 263.759C646.385 264.914 649.215 267.105 651.329 270.816L652.984 273.721L654.054 270.553C655.166 267.259 657.497 262.791 660.791 259.974C662.419 258.581 664.228 257.638 666.199 257.375C668.153 257.115 670.409 257.503 672.993 259.015L673.729 259.445L674.476 259.033Z"
-            strokeWidth={3}
-          />
-          <path d="M786.5 365.5V454.5H47.5V365.5H786.5Z" strokeWidth={3} />
-          <path
-            d="M821.781 233.257L775.517 233.257C769.718 233.257 765.017 237.958 765.017 243.757L765.017 454.901C765.017 460.7 769.718 465.401 775.517 465.401L821.781 465.401C827.58 465.401 832.281 460.7 832.281 454.901L832.281 243.757C832.281 237.958 827.58 233.257 821.781 233.257Z"
-            strokeWidth={3}
-          />
-          <rect
-            height="232.145"
-            rx="10.5"
-            strokeWidth={3}
-            width="67.2615"
-            x="17.5"
-            y="229.292"
-          />
-          <path d="M253.376 427.317L253.704 425.853H253.704L253.376 427.317ZM245.915 420.19L247.363 419.796L245.915 420.19ZM245.947 414.605C245.729 413.806 244.904 413.335 244.105 413.553C243.306 413.771 242.835 414.595 243.053 415.395L245.947 414.605ZM270.588 424.834L270.97 426.284L270.588 424.834ZM305.818 415.227L306.219 416.672L305.818 415.227ZM308.789 400.273L308.039 398.974L308.789 400.273ZM313.112 397.778L313.862 399.077L313.112 397.778ZM321.555 397.778L322.305 396.479V396.479L321.555 397.778ZM331.468 403.501L330.718 404.8L331.468 403.501ZM346.95 399.352L348.249 400.102L348.249 400.102L346.95 399.352ZM347.418 398.541L346.119 397.791L347.418 398.541ZM368.563 391.212L369.624 392.272L368.563 391.212ZM375.013 384.761L376.074 385.822L376.074 385.822L375.013 384.761ZM380.38 380.643L379.63 379.344L380.38 380.643ZM358.794 393.829L359.182 392.38L358.794 393.829ZM307.834 407.531L308.894 406.47L307.834 407.531ZM253.539 434.414L254.6 435.475H254.6L253.539 434.414ZM261.16 429.062L261.488 427.598L261 427.489L260.672 428.953L260.344 430.416L260.832 430.526L261.16 429.062ZM260.672 428.953L261 427.489L253.704 425.853L253.376 427.317L253.048 428.781L260.344 430.416L260.672 428.953ZM245.915 420.19L247.363 419.796L245.947 414.605L244.5 415L243.053 415.395L244.468 420.585L245.915 420.19ZM253.376 427.317L253.704 425.853C250.631 425.165 248.191 422.834 247.363 419.796L245.915 420.19L244.468 420.585C245.589 424.695 248.891 427.849 253.048 428.781L253.376 427.317ZM270.588 424.834L270.97 426.284C280.46 423.785 294.937 419.805 306.219 416.672L305.818 415.227L305.416 413.782C294.135 416.914 279.675 420.889 270.206 423.383L270.588 424.834ZM308.789 400.273L309.539 401.572L313.862 399.077L313.112 397.778L312.362 396.479L308.039 398.974L308.789 400.273ZM321.555 397.778L320.805 399.077L330.718 404.8L331.468 403.501L332.218 402.202L322.305 396.479L321.555 397.778ZM346.95 399.352L348.249 400.102L348.717 399.291L347.418 398.541L346.119 397.791L345.651 398.602L346.95 399.352ZM368.563 391.212L369.624 392.272L376.074 385.822L375.013 384.761L373.953 383.701L367.502 390.151L368.563 391.212ZM380.38 380.643L381.13 381.942L387.931 378.016L387.181 376.717L386.431 375.418L379.63 379.344L380.38 380.643ZM375.013 384.761L376.074 385.822C377.583 384.313 379.282 383.009 381.13 381.942L380.38 380.643L379.63 379.344C377.555 380.542 375.647 382.007 373.953 383.701L375.013 384.761ZM358.794 393.829L358.406 395.278C362.414 396.352 366.69 395.206 369.624 392.272L368.563 391.212L367.502 390.151C365.326 392.327 362.155 393.177 359.182 392.38L358.794 393.829ZM347.418 398.541L348.717 399.291C350.667 395.915 354.64 394.269 358.406 395.278L358.794 393.829L359.182 392.38C354.105 391.02 348.747 393.239 346.119 397.791L347.418 398.541ZM331.468 403.501L330.718 404.8C336.856 408.344 344.705 406.241 348.249 400.102L346.95 399.352L345.651 398.602C342.935 403.306 336.921 404.917 332.218 402.202L331.468 403.501ZM313.112 397.778L313.862 399.077C316.01 397.836 318.656 397.836 320.805 399.077L321.555 397.778L322.305 396.479C319.228 394.702 315.438 394.702 312.362 396.479L313.112 397.778ZM307.834 407.531L308.894 406.47C307.447 405.022 307.766 402.596 309.539 401.572L308.789 400.273L308.039 398.974C304.558 400.984 303.931 405.749 306.773 408.591L307.834 407.531ZM305.818 415.227L306.219 416.672C310.732 415.419 312.198 409.774 308.894 406.47L307.834 407.531L306.773 408.591C308.46 410.279 307.705 413.146 305.416 413.782L305.818 415.227ZM270.588 424.834L270.206 423.383C266.654 424.318 263.146 425.765 259.913 427.658L260.672 428.953L261.43 430.247C264.424 428.492 267.679 427.151 270.97 426.284L270.588 424.834ZM260.672 428.953L259.913 427.658C257.202 429.247 254.676 431.157 252.479 433.354L253.539 434.414L254.6 435.475C256.6 433.475 258.919 431.718 261.43 430.247L260.672 428.953ZM231.984 455.969L233.045 457.03L254.6 435.475L253.539 434.414L252.479 433.354L230.924 454.908L231.984 455.969Z" />
-          <path
-            d="M136 486.083C136 480.515 137.08 476 138.411 476C139.743 476 140.822 480.515 140.822 486.083C140.822 491.652 141.902 496.167 143.234 496.167C144.565 496.167 145.645 491.652 145.645 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M140.822 486.083C140.822 480.515 141.902 476 143.233 476C144.565 476 145.645 480.515 145.645 486.083C145.645 491.652 146.724 496.167 148.056 496.167C149.388 496.167 150.467 491.652 150.467 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M145.645 486.083C145.645 480.514 146.724 476 148.056 476C149.387 476 150.467 480.514 150.467 486.083C150.467 491.652 151.547 496.167 152.878 496.167C154.21 496.167 155.289 491.652 155.289 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M150.469 486.083C150.469 480.514 151.548 476 152.88 476C154.212 476 155.291 480.514 155.291 486.083C155.291 491.652 156.371 496.167 157.702 496.167C159.034 496.167 160.114 491.652 160.114 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M155.291 486.083C155.291 480.514 156.371 476 157.702 476C159.034 476 160.113 480.514 160.113 486.083C160.113 491.652 161.193 496.167 162.525 496.167C163.856 496.167 164.936 491.652 164.936 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M160.113 486.083C160.113 480.514 161.193 476 162.524 476C163.856 476 164.936 480.514 164.936 486.083C164.936 491.652 166.015 496.167 167.347 496.167C168.679 496.167 169.758 491.652 169.758 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M164.936 486.083C164.936 480.514 166.015 476 167.347 476C168.678 476 169.758 480.514 169.758 486.083C169.758 491.652 170.838 496.167 172.169 496.167C173.501 496.167 174.58 491.652 174.58 486.083"
-            strokeWidth={1.5}
-          />
-          <path
-            d="M169.758 486.083C169.758 480.514 170.837 476 172.169 476C173.501 476 174.58 480.514 174.58 486.083C174.58 491.652 175.66 496.167 176.991 496.167C178.323 496.167 179.403 491.652 179.403 486.083"
-            strokeWidth={1.5}
-          />
-        </svg>
-      </div>
-
-      <h1 className="font-outfit text-center text-2xl font-semibold md:text-4xl">
-        Well!, this page is as broken as our couch!
-      </h1>
-      <p className="my-0 text-center text-black/60 dark:text-white md:my-1.5">
-        Missing Page, but your next favourite chair is just a click away.
+    <div className="flex flex-col items-center justify-center text-center min-h-screen bg-white px-4 font-sans">
+      <img
+        src="https://cdn-icons-png.flaticon.com/512/7486/7486758.png"
+        alt="Error"
+        className="w-72 mb-6"
+      />
+      <h2 className="text-2xl md:text-3xl font-semibold text-gray-500 mb-2">
+        404, Halaman Tidak Ditemukan
+      </h2>
+      <p className="text-gray-500 max-w-lg mb-8 leading-relaxed">
+        Terjadi kesalahan. Tampaknya permintaan Anda tidak dapat ditemukan.
+        Kemungkinan tautannya rusak atau halamannya telah dihapus.
       </p>
-      <EventButton buttonName="Back to Home" redirect="/" />
+      <div className="flex space-x-4">
+        <button
+          onClick={handleGoBack}
+          className="bg-orange-500 hover:bg-orange-600 text-white font-medium px-6 py-2 rounded-md transition-colors"
+        >
+          ← KEMBALI
+        </button>
+        <button
+          onClick={handleGoHome}
+          className="bg-orange-100 hover:bg-orange-200 text-orange-600 border border-orange-400 font-medium px-6 py-2 rounded-md transition-colors"
+        >
+          ⟳ KEMBALI KE BERANDA
+        </button>
+      </div>
     </div>
   );
 }
+
